% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc_cor.R
\name{my_calc_cor}
\alias{my_calc_cor}
\title{Calculate the correlation for two kinds of score}
\usage{
my_calc_cor(data_1, data_2, score = "each_row")
}
\arguments{
\item{data_1}{tibble/matrix}

\item{data_2}{tibble/matrix}

\item{score}{string, "each_row" - score by row, "each_col" - score by col}
}
\value{
matrix
}
\description{
Calculate the correlation for two kinds of score
}
