% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc_purity.R
\name{my_calc_purity}
\alias{my_calc_purity}
\title{Calculate the tumor purity using package estimate.}
\usage{
my_calc_purity(data)
}
\arguments{
\item{data}{tibble, feature by row and case by col}
}
\value{
matrix
}
\description{
Calculate the tumor purity using package estimate.
}
