% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_diff_wilcox.R
\name{my_diff_wilcox}
\alias{my_diff_wilcox}
\title{Wilcoxon ranksum method for diff analysis}
\usage{
my_diff_wilcox(
  data,
  group_info,
  adjust.method = "fdr",
  data_type = "gene",
  p.value.cutoff = 1,
  fold.change.cutoff = 1,
  FC_method = mean
)
}
\arguments{
\item{data}{tibble, feature by row and case by col}

\item{group_info}{tibble, with two col, first is case_id and second is group_name}

\item{data_type}{string, "gene" or other}

\item{FC_method}{}
}
\value{
tibble
}
\description{
Wilcoxon ranksum method for diff analysis
}
