% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_enrichment_DEG.R
\name{my_enrichment_DEG}
\alias{my_enrichment_DEG}
\title{Enrichment analysis using clusterProfiler.}
\usage{
my_enrichment_DEG(
  DEG_list,
  DB = "GO",
  sub_DB = "BP",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2
)
}
\arguments{
\item{DEG_list}{vector, the Diff Expr Genes list}

\item{DB}{string, "GO", "KEGG", "All"}

\item{sub_DB}{string, "BP", "MF", "CC", "All"}

\item{qvalueCutoff}{}
}
\value{
list
}
\description{
Enrichment analysis using clusterProfiler.
}
