% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_get_geneset_genes.R
\name{my_get_geneset_genes}
\alias{my_get_geneset_genes}
\title{Get the genes of the genesets}
\usage{
my_get_geneset_genes(geneset_file)
}
\arguments{
\item{geneset_file}{string, the file of geneset}
}
\value{
list
}
\description{
Get the genes of the genesets
}
