% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_get_tibble_row.R
\name{my_get_tibble_row}
\alias{my_get_tibble_row}
\title{Get the row data of tibble}
\usage{
my_get_tibble_row(data, row, by = "index", keep_first = FALSE)
}
\arguments{
\item{data}{tibble}

\item{row}{integer/string}

\item{by}{string, "index" - by index, "name" - by rowname}

\item{keep_first}{boolean, wheather to keep first col}
}
\value{
vector
}
\description{
Get the row data of tibble
}
