% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{my_GSEA}
\alias{my_GSEA}
\title{GSEA analysis using clusterProfiler or fgsea}
\usage{
my_GSEA(
  genes,
  logFC,
  method = "gesGO",
  sub_DB = "MF",
  pvalueCutoff = 0.05,
  gmt_dir = NULL
)
}
\arguments{
\item{genes}{vector,logFC and genes should be tha same order}

\item{logFC}{vector}

\item{method}{string, "gesGO" or "fgsea"}

\item{sub_DB}{string, "MF","BP","CC", "GO_All", "KEGG", and "GO"("GO" is only for fgsea)}

\item{gmt_dir}{string, the gmt dir to perform special geneset analysis}
}
\value{
list
}
\description{
GSEA analysis using clusterProfiler or fgsea
}
