% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA_plot_load.R
\name{my_GSEA_plot_load}
\alias{my_GSEA_plot_load}
\title{Load the GSEA plot}
\usage{
my_GSEA_plot_load(data, geneset)
}
\arguments{
\item{data}{list, which generated from my_load_GSEA}

\item{geneset}{string, spcify the geneset to load}
}
\value{
image object, which should be displayed using package EBImage
}
\description{
Load the GSEA plot
}
