% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_inner_join_col.R
\name{my_inner_join_col}
\alias{my_inner_join_col}
\title{The col version of inner_join}
\usage{
my_inner_join_col(data_1, data_2)
}
\arguments{
\item{data_1}{tibble}

\item{data_2}{tibble}
}
\value{
tibble
}
\description{
The col version of inner_join
}
