% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_intersect_GSEA.R
\name{my_intersect_GSEA}
\alias{my_intersect_GSEA}
\title{Get the intersect GSEA results of two datasets.}
\usage{
my_intersect_GSEA(data_1, data_2, FDR.cutoff = 1, p.cutoff = 0.5)
}
\arguments{
\item{data_1}{list, which generated from my_load_GSEA}

\item{data_2}{list, which generated from my_load_GSEA}

\item{p.cutoff}{}
}
\value{
list, with two tibble, for pos and neg
}
\description{
Get the intersect GSEA results of two datasets.
}
