% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_limma.R
\name{my_limma}
\alias{my_limma}
\title{limma method to get DEG}
\usage{
my_limma(
  data,
  group_info,
  adjust.method = "fdr",
  data_type = "counts",
  p.value.cutoff = 1,
  fold.change.cutoff = 1,
  filterLow = -100
)
}
\arguments{
\item{data}{tibble, gene by row and case by col}

\item{group_info}{tibble, with two col, first is case_id and second is group_name}

\item{data_type}{string, "counts" or "score"}

\item{filterLow}{}
}
\value{
tibble.
}
\description{
limma method to get DEG
}
