% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_load_GSEA.R
\name{my_load_GSEA}
\alias{my_load_GSEA}
\title{Load GSEA results from given dir}
\usage{
my_load_GSEA(dir, pos_name, neg_name)
}
\arguments{
\item{dir}{string, the directory of the results}

\item{pos_name}{string, the name of POS group, such as "sensitive"}

\item{neg_name}{string, the name of NEG group, such as "resistant"}
}
\value{
list, with dir, id, pos_results, neg_results
}
\description{
Load GSEA results from given dir
}
