% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_load_tibble.R
\name{my_load_tibble}
\alias{my_load_tibble}
\title{Load tibble from file}
\usage{
my_load_tibble(file, type = NULL, sheetName = NULL)
}
\arguments{
\item{file}{string, filename}

\item{type}{string, "csv", "tsv", "txt", "xlsx"}

\item{sheetName}{string}
}
\value{
tibble
}
\description{
Load tibble from file
}
