% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_match_both_rowcol_inner.R
\name{my_match_both_rowcol_inner}
\alias{my_match_both_rowcol_inner}
\title{Re-arrange the row and col to make the order of the two data is same
The dismatch row or col would be drop.}
\usage{
my_match_both_rowcol_inner(data_1, data_2)
}
\arguments{
\item{data_1}{tibble}

\item{data_2}{tibble}
}
\value{
list, including two tibble
}
\description{
Re-arrange the row and col to make the order of the two data is same
The dismatch row or col would be drop.
}
