% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_match_col.R
\name{my_match_col}
\alias{my_match_col}
\title{Re-arrange the col to make the order of the two data is same}
\usage{
my_match_col(data_1, data_2, method = "inner")
}
\arguments{
\item{data_1}{tibble}

\item{data_2}{tibble}

\item{method}{string, "inner" - like inner join, "left" - only keep the col of data_1, "right" - only keep the col of data_2}
}
\value{
list, including two tibble
}
\description{
Re-arrange the col to make the order of the two data is same
}
