% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_match_row.R
\name{my_match_row}
\alias{my_match_row}
\title{Re-arrange the row to make the order of the two data is same}
\usage{
my_match_row(data_1, data_2, by_1 = 1, by_2 = 1, method = "inner")
}
\arguments{
\item{data_1}{tibble}

\item{data_2}{tibble}

\item{method}{string, "inner" - like inner join, "left" - only keep the row of data_1, "right" - only keep the row of data_2}
}
\value{
list, including two tibble
}
\description{
Re-arrange the row to make the order of the two data is same
}
