% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_prepare_GSEA.R
\name{my_prepare_GSEA}
\alias{my_prepare_GSEA}
\title{Generate the data for GSEA}
\usage{
my_prepare_GSEA(data, group_info, name)
}
\arguments{
\item{data}{tibble, feature by row and case by col}

\item{group_info}{tibble, with two col, first is case_id and second is group_name}

\item{name}{string, which is used to name the file}
}
\description{
Generate the data for GSEA
}
