% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_preprocess_RNA.R
\name{my_preprocess_RNA}
\alias{my_preprocess_RNA}
\title{Preprocess RNA data}
\usage{
my_preprocess_RNA(
  data,
  correctNames = TRUE,
  keep_duplicate = FALSE,
  combine_method = median,
  filterLow = 0.001
)
}
\arguments{
\item{data}{tibble}

\item{filterLow}{}
}
\value{
tibble
}
\description{
Preprocess RNA data
}
