% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_save_tibble.R
\name{my_save_tibble}
\alias{my_save_tibble}
\title{Save tibble to file}
\usage{
my_save_tibble(data, file, type = NULL, sheetName = "Sheet1")
}
\arguments{
\item{data}{tibble}

\item{file}{string,filename}

\item{type}{string, "csv", "tsv", "txt", "xlsx"}

\item{sheetName}{string}
}
\description{
Save tibble to file
}
