% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_simple_volcanoPlot.R
\name{my_simple_volcanoPlot}
\alias{my_simple_volcanoPlot}
\title{Get simple volcano plot using base plot}
\usage{
my_simple_volcanoPlot(
  logFC,
  pvalue,
  logFC.cutoff = 1,
  pvalue.cutoff = 0.05,
  xlab = "Log2(Fold change)",
  ylab = "-Log10(Pvalue)",
  main = "",
  xlim = NULL,
  ylim = NULL,
  linetype = "solid"
)
}
\arguments{
\item{logFC}{vector, logFC and pvalue should be tha same order}

\item{pvalue}{vector}

\item{linetype}{}
}
\value{
plot
}
\description{
Get simple volcano plot using base plot
}
