% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary_tibble.R
\name{my_summary_tibble}
\alias{my_summary_tibble}
\title{Extended summary for tibble}
\usage{
my_summary_tibble(data, fun)
}
\arguments{
\item{data}{tibble}

\item{fun}{function}
}
\value{
vector
}
\description{
Extended summary for tibble
}
