% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary_tibble_row.R
\name{my_summary_tibble_row}
\alias{my_summary_tibble_row}
\title{Extended summary for tibble row}
\usage{
my_summary_tibble_row(data, fun)
}
\arguments{
\item{data}{tibble}

\item{fun}{function}
}
\value{
vector
}
\description{
Extended summary for tibble row
}
