% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_tibble.R
\name{my_t_tibble}
\alias{my_t_tibble}
\title{Transpose for tibble}
\usage{
my_t_tibble(data, new_name_col_1)
}
\arguments{
\item{data}{tibble}

\item{new_name_col_1}{string, the name for col_1 of result}
}
\value{
tibble
}
\description{
Transpose for tibble
}
