% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tibble2matrix.R
\name{my_tibble2matrix}
\alias{my_tibble2matrix}
\title{transform tibble to matrix with rownames}
\usage{
my_tibble2matrix(data)
}
\arguments{
\item{data}{tibble}
}
\value{
matrix
}
\description{
transform tibble to matrix with rownames
}
