% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_volcanoPlot.R
\name{my_volcanoPlot}
\alias{my_volcanoPlot}
\title{Get the volcano plot using ggplot2}
\usage{
my_volcanoPlot(
  logFC,
  pvalue,
  logFC.cutoff = 1,
  pvalue.cutoff = 0.05,
  xlab = "Log2(Fold change)",
  ylab = "-Log10(Pvalue)",
  main = "",
  xlim = NULL,
  ylim = NULL,
  linetype = "solid",
  genes = NULL,
  highlight_genes = NULL
)
}
\arguments{
\item{logFC}{vector, logFC, pvalue and genes should be tha same order}

\item{pvalue}{vector}

\item{genes}{vector}

\item{highlight_genes}{the genes to be highlight with label}
}
\value{
ggplot object.
}
\description{
Get the volcano plot using ggplot2
}
